chore: switch to nomo repo
