+++ .npmrc
@@ -1,0 +1,2 @@
+prefer-workspace-packages=false
+save-workspace-protocol=false
