+++ packages/main/forge.config.js
