+++ packages/main/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@croffle/main",
+  "version": "1.0.0",
+  "description": "Project Croffle Electron Main Process",
+  "private": true,
+  "main": "dist/main.js",
+  "scripts": {
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make",
+    "dev": "tsc && cross-env NODE_ENV=development npx electron@38.1.2 .",
+    "build": "tsc"
+  },
+  "devDependencies": {
+    "@electron-forge/cli": "^7.9.0",
+    "@electron-forge/maker-deb": "^7.9.0",
+    "@electron-forge/maker-rpm": "^7.9.0",
+    "@electron-forge/maker-squirrel": "^7.9.0",
+    "@electron-forge/maker-zip": "^7.9.0",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.9.0",
+    "@electron-forge/plugin-fuses": "^7.9.0",
+    "@electron/fuses": "^1.8.0",
+    "@types/node": "^24.3.1",
+    "electron": "^38.1.2"
+  },
+  "dependencies": {
+    "electron-squirrel-startup": "^1.0.1"
+  }
+}
