+++ packages/main/src/main.ts
@@ -1,0 +1,42 @@
+import { ipcHandlers } from "./ipc/ipcHandle";
+import { app, BrowserWindow } from "electron";
+import * as path from "path";
+
+const isDev = process.env.NODE_ENV === "development";
+
+function createWindow(): void {
+  const mainWindow = new BrowserWindow({
+    width: 1200,
+    height: 800,
+    webPreferences: {
+      nodeIntegration: false,
+      contextIsolation: true,
+      preload: path.join(__dirname, "./electron/preload.js"),
+    },
+  });
+
+  if (isDev) {
+    mainWindow.loadURL("http://localhost:5173");
+    mainWindow.webContents.openDevTools();
+  } else {
+    mainWindow.loadFile(path.join(__dirname, "../../renderer/dist/index.html"));
+  }
+}
+
+ipcHandlers();
+
+app.whenReady().then(() => {
+  createWindow();
+
+  app.on("activate", () => {
+    if (BrowserWindow.getAllWindows().length === 0) {
+      createWindow();
+    }
+  });
+});
+
+app.on("window-all-closed", () => {
+  if (process.platform !== "darwin") {
+    app.quit();
+  }
+});
