+++ packages/renderer/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@croffle/renderer",
+  "version": "1.0.0",
+  "description": "Project Croffle Vue Renderer Process",
+  "private": true,
+  "main": "src/main.ts",
+  "scripts": {
+    "dev": "vite",
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^6.0.1",
+    "@vue/tsconfig": "^0.8.1",
+    "prettier-plugin-tailwindcss": "^0.6.14",
+    "tailwindcss": "^4.1.13",
+    "vite": "^7.1.5",
+    "vue": "^3.5.21",
+    "vue-eslint-parser": "^10.2.0",
+    "vue-tsc": "^3.0.6"
+  },
+  "dependencies": {
+    "pinia": "^3.0.3",
+    "vue": "^3.5.21",
+    "vue-router": "^4.5.1"
+  }
+}
