+++ packages/renderer/src/main.ts
@@ -1,0 +1,7 @@
+import { createApp } from 'vue';
+import App from './App.vue';
+import './style.css';
+
+const app = createApp(App);
+
+app.mount('#app');
