+++ packages/renderer/src/style.css
@@ -1,0 +1,3 @@
+@import 'tailwindcss';
+
+@custom-variant dark (&:where([data-theme=dark], [data-theme=dark] *));
