--- src/electron/preload.ts
@@ -1,7 +1,0 @@
-import { contextBridge, ipcRenderer } from "electron";
-
-contextBridge.exposeInMainWorld("electronAPI", {
-  minimize: () => ipcRenderer.invoke("window:minimize"),
-  maximize: () => ipcRenderer.invoke("window:maximize"),
-  close: () => ipcRenderer.invoke("window:close"),
-});
