--- src/ipc/ipcHandle.ts
@@ -1,26 +1,0 @@
-import { BrowserWindow, ipcMain } from "electron";
-
-export function ipcHandlers() {
-  ipcMain.handle("window:minimize", (event) => {
-    const window = BrowserWindow.fromWebContents(event.sender);
-    if (window) window.minimize();
-  });
-
-  ipcMain.handle("window:maximize", (event) => {
-    const window = BrowserWindow.fromWebContents(event.sender);
-    if (window) {
-      if (window.isMaximized()) {
-        window.unmaximize();
-      } else {
-        window.maximize();
-      }
-    }
-  });
-
-  ipcMain.handle("window:close", (event) => {
-    const window = BrowserWindow.fromWebContents(event.sender);
-    if (window) window.close();
-  });
-
-  // Handle IPC events here
-}
